+++ pages/home/transaksi/nko.vue
@@ -266,6 +266,7 @@ const detailItems = computed<DetailDataItem[]>(() => {
       :subtitle="modalSubtitle"
       :sections="formSections"
       :submitting="submitting"
+      draft-key="transaksi-nko"
       @submit="handleSubmit"
       @cancel="modalOpen = false"
     />