+++ pages/home/transaksi/prognosa.vue
@@ -288,6 +288,7 @@ const detailItems = computed<DetailDataItem[]>(() => {
       :subtitle="modalSubtitle"
       :sections="formSections"
       :submitting="submitting"
+      draft-key="transaksi-prognosa"
       @submit="handleSubmit"
       @cancel="modalOpen = false"
     />